requirements.txt
Flask==0.11.1

# Application server for both development and production.
gunicorn==19.4.5

# Testing and static analysis.
pytest==3.0.4
pytest-cov==2.4.0
flake8==3.2.1

# CLI commands.
click==6.6

# Data and workers.
psycopg2==2.6.2
Flask-SQLAlchemy==2.1
SQLAlchemy-Utils==0.32.11
redis==2.10.5
celery==4.0.0

# Forms.
Flask-WTF==0.13.1
WTForms-Components==0.10.1

# Flask Extensions.
flask-debugtoolbar==0.10.0
Flask-Login==0.4.0
Flask-RESTful==0.3.5
flask-marshmallow==0.7.0